Implement graceful shutdown of start command via CancellationToken (#119)
The start command internally spawns multiple async tasks using
`tokio::spawn`, without returning their corresponding join handles.
This means there is no way of terminating these async tasks.

For the CLI this is not a problem, because the server simply runs until
the entire program is aborted using SIGINT (CTRL+C).

For testing, the `sync` task automatically exits once the wiremock
server stops responding, which happens when it gets dropped at the end
of the test. However, future use cases such as serving telemetry, which
is done in an endless loop, require an explicit mechanism for
terminating tasks in tests.

To this end, the `start` command now receives a `CancellationToken`
which can be used to cancel nested tasks. The main thread now intercepts
CTRL+C to signal the token and print a shutdown message. This mechanism
can be used to implement additional shutdown logic in the future if
needed.